Refactor Search.java & UserService.java
